requirements.txt
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.3
psycopg2-binary
django-filter
markdown
djangorestframework-gis
drf-extra-fields
gpxpy
